requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
font-hanken-grotesk==0.0.2
font-intuitive==0.0.4
idna==3.4
inky==1.5.0
numpy==1.25.0
Pillow==9.5.0
ply==3.11
pyasn1==0.5.0
pycryptodomex==3.18.0
pysmi==0.3.4
pysnmp==4.4.12
python-dotenv==1.0.0
requests==2.31.0
smbus2==0.4.2
spidev==3.6
urllib3==2.0.3
